This now compiles and passes tests under CLANG
