+++ include/simdjson/stage1_find_marks.h
@@ -3,99 +3,14 @@
 
 #include <cassert>
 #include "simdjson/common_defs.h"
+#include "simdjson/simdjson.h"
 #include "simdjson/parsedjson.h"
 #include "simdjson/portability.h"
 
-#ifdef IS_X86_64
-#include "simdjson/simdutf8check.h"
-#elif IS_ARM64
-#include "simdjson/simdutf8check_neon.h"
-#endif
-
-//#define TRANSPOSE
-
 namespace simdjson {
+
 template<architecture>
 struct simd_input;
-
-#ifdef IS_X86_64
-TARGET_HASWELL
-template<>
-struct simd_input<architecture::haswell>
-{
-  __m256i lo;
-  __m256i hi;
-};
-UNTARGET_REGION
-
-TARGET_WESTMERE
-template<>
-struct simd_input<architecture::westmere>
-{
-  __m128i v0;
-  __m128i v1;
-  __m128i v2;
-  __m128i v3;
-};
-UNTARGET_REGION
-#endif // IS_X86_64
-
-#ifdef IS_ARM64
-template<> struct simd_input<architecture::arm64>
-{
-#ifndef TRANSPOSE
-  uint8x16_t i0;
-  uint8x16_t i1;
-  uint8x16_t i2;
-  uint8x16_t i3;
-#else
-  uint8x16x4_t i;
-#endif
-};
-
-really_inline
-uint16_t neonmovemask(uint8x16_t input) {
-  const uint8x16_t bitmask = { 0x01, 0x02, 0x4, 0x8, 0x10, 0x20, 0x40, 0x80,
-                               0x01, 0x02, 0x4, 0x8, 0x10, 0x20, 0x40, 0x80};
-  uint8x16_t minput = vandq_u8(input, bitmask);
-  uint8x16_t tmp = vpaddq_u8(minput, minput);
-  tmp = vpaddq_u8(tmp, tmp);
-  tmp = vpaddq_u8(tmp, tmp);
-  return vgetq_lane_u16(vreinterpretq_u16_u8(tmp), 0);
-}
-
-really_inline
-uint64_t neonmovemask_bulk(uint8x16_t p0, uint8x16_t p1, uint8x16_t p2, uint8x16_t p3) {
-#ifndef TRANSPOSE
-  const uint8x16_t bitmask = { 0x01, 0x02, 0x4, 0x8, 0x10, 0x20, 0x40, 0x80,
-                               0x01, 0x02, 0x4, 0x8, 0x10, 0x20, 0x40, 0x80};
-  uint8x16_t t0 = vandq_u8(p0, bitmask);
-  uint8x16_t t1 = vandq_u8(p1, bitmask);
-  uint8x16_t t2 = vandq_u8(p2, bitmask);
-  uint8x16_t t3 = vandq_u8(p3, bitmask);
-  uint8x16_t sum0 = vpaddq_u8(t0, t1);
-  uint8x16_t sum1 = vpaddq_u8(t2, t3);
-  sum0 = vpaddq_u8(sum0, sum1);
-  sum0 = vpaddq_u8(sum0, sum0);
-  return vgetq_lane_u64(vreinterpretq_u64_u8(sum0), 0);
-#else
-  const uint8x16_t bitmask1 = { 0x01, 0x10, 0x01, 0x10, 0x01, 0x10, 0x01, 0x10,
-                                0x01, 0x10, 0x01, 0x10, 0x01, 0x10, 0x01, 0x10};
-  const uint8x16_t bitmask2 = { 0x02, 0x20, 0x02, 0x20, 0x02, 0x20, 0x02, 0x20,
-                                0x02, 0x20, 0x02, 0x20, 0x02, 0x20, 0x02, 0x20};
-  const uint8x16_t bitmask3 = { 0x04, 0x40, 0x04, 0x40, 0x04, 0x40, 0x04, 0x40,
-                                0x04, 0x40, 0x04, 0x40, 0x04, 0x40, 0x04, 0x40};
-  const uint8x16_t bitmask4 = { 0x08, 0x80, 0x08, 0x80, 0x08, 0x80, 0x08, 0x80,
-                                0x08, 0x80, 0x08, 0x80, 0x08, 0x80, 0x08, 0x80};
-  uint8x16_t t0 = vandq_u8(p0, bitmask1);
-  uint8x16_t t1 = vbslq_u8(bitmask2, p1, t0);
-  uint8x16_t t2 = vbslq_u8(bitmask3, p2, t1);
-  uint8x16_t tmp = vbslq_u8(bitmask4, p3, t2);
-  uint8x16_t sum = vpaddq_u8(tmp, tmp);
-  return vgetq_lane_u64(vreinterpretq_u64_u8(sum), 0);
-#endif
-}
-#endif // IS_ARM64
 
 template<architecture T>
 uint64_t compute_quote_mask(uint64_t quote_bits);
@@ -113,495 +28,40 @@ namespace {
   }
 }
 
-// In practice, if you have NEON or __PCLMUL__, you would
-// always want to use them, but it might be useful, for research
-// purposes, to disable it willingly, that's what SIMDJSON_AVOID_CLMUL
-// does.
-// Also: we don't know of an instance where AVX2 is supported but 
-// where clmul is not supported, so check for both, to be sure.
-#ifdef SIMDJSON_AVOID_CLMUL
-template<architecture T> really_inline
-uint64_t compute_quote_mask(uint64_t quote_bits) {
-  return portable_compute_quote_mask(quote_bits);
-}
-#else
-template<architecture>
-uint64_t compute_quote_mask(uint64_t quote_bits);
-
-#ifdef IS_X86_64
-TARGET_HASWELL
-template<> really_inline
-uint64_t compute_quote_mask<architecture::haswell>(uint64_t quote_bits) {
-  // There should be no such thing with a processing supporting avx2
-  // but not clmul.
-  uint64_t quote_mask = _mm_cvtsi128_si64(_mm_clmulepi64_si128(
-      _mm_set_epi64x(0ULL, quote_bits), _mm_set1_epi8(0xFF), 0));
-  return quote_mask;
-}
-UNTARGET_REGION
-
-TARGET_WESTMERE
-template<> really_inline
-uint64_t compute_quote_mask<architecture::westmere>(uint64_t quote_bits) {
-  return _mm_cvtsi128_si64(_mm_clmulepi64_si128(
-      _mm_set_epi64x(0ULL, quote_bits), _mm_set1_epi8(0xFF), 0));
-}
-UNTARGET_REGION
-#endif //IS_X86_64
-
-#ifdef IS_ARM64
-template<> really_inline
-uint64_t compute_quote_mask<architecture::arm64>(uint64_t quote_bits) {
-#ifdef __ARM_FEATURE_CRYPTO // some ARM processors lack this extension
-  return vmull_p64( -1ULL, quote_bits);
-#else
-  return portable_compute_quote_mask(quote_bits);
-#endif 
-}
-#endif // IS_ARM64
-#endif // SIMDJSON_AVOID_CLMUL
-
 // Holds the state required to perform check_utf8().
 template<architecture>
 struct utf8_checking_state;
 
-#ifdef IS_X86_64
-#include "simdjson/simdutf8check.h"
-TARGET_HASWELL
-template<>
-struct utf8_checking_state<architecture::haswell>
-{
-  __m256i has_error;
-  avx_processed_utf_bytes previous;
-  utf8_checking_state() {
-    has_error = _mm256_setzero_si256();
-    previous.rawbytes = _mm256_setzero_si256();
-    previous.high_nibbles = _mm256_setzero_si256();
-    previous.carried_continuations =_mm256_setzero_si256();
-  }
-};
-UNTARGET_REGION
-
-TARGET_WESTMERE
-template<>
-struct utf8_checking_state<architecture::westmere>
-{
-  __m128i has_error = _mm_setzero_si128();
-  processed_utf_bytes previous {
-    _mm_setzero_si128(), // rawbytes
-    _mm_setzero_si128(), // high_nibbles
-    _mm_setzero_si128()  // carried_continuations
-  };
-};
-UNTARGET_REGION
-#endif
-
-#ifdef IS_ARM64
-template<>
-struct utf8_checking_state<architecture::arm64>
-{
-  int8x16_t has_error {};
-  processed_utf_bytes previous {};
-};
-
-// Checks that all bytes are ascii
-really_inline
-bool check_ascii_neon(simd_input<architecture::arm64> in) {
-  // checking if the most significant bit is always equal to 0.
-  uint8x16_t highbit = vdupq_n_u8(0x80);
-  uint8x16_t t0 = vorrq_u8(in.i0, in.i1);
-  uint8x16_t t1 = vorrq_u8(in.i2, in.i3);
-  uint8x16_t t3 = vorrq_u8(t0, t1);
-  uint8x16_t t4 = vandq_u8(t3, highbit);
-  uint64x2_t v64 = vreinterpretq_u64_u8(t4);
-  uint32x2_t v32 = vqmovn_u64(v64);
-  uint64x1_t result = vreinterpret_u64_u32(v32);
-  return vget_lane_u64(result, 0) == 0;
-}
-#endif
 
 template<architecture T>
 void check_utf8(simd_input<T> in, utf8_checking_state<T>& state);
-
-#ifdef IS_X86_64
-TARGET_HASWELL
-template<> really_inline
-void check_utf8<architecture::haswell>(simd_input<architecture::haswell> in,
-                utf8_checking_state<architecture::haswell>& state) {
-  __m256i highbit = _mm256_set1_epi8(0x80);
-  if ((_mm256_testz_si256(_mm256_or_si256(in.lo, in.hi), highbit)) == 1) {
-    // it is ascii, we just check continuation
-    state.has_error = _mm256_or_si256(
-        _mm256_cmpgt_epi8(
-            state.previous.carried_continuations,
-            _mm256_setr_epi8(9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9,
-                             9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 1)),
-        state.has_error);
-  } else {
-    // it is not ascii so we have to do heavy work
-    state.previous = avxcheckUTF8Bytes(in.lo, &(state.previous), &(state.has_error));
-    state.previous = avxcheckUTF8Bytes(in.hi, &(state.previous), &(state.has_error));
-  }
-}
-UNTARGET_REGION
-
-TARGET_WESTMERE
-template<> really_inline
-void check_utf8<architecture::westmere>(simd_input<architecture::westmere> in,
-                utf8_checking_state<architecture::westmere>& state) {
-  __m128i highbit = _mm_set1_epi8(0x80);
-  if ((_mm_testz_si128(_mm_or_si128(in.v0, in.v1), highbit)) == 1) {
-    // it is ascii, we just check continuation
-    state.has_error = _mm_or_si128(
-        _mm_cmpgt_epi8(
-            state.previous.carried_continuations,
-            _mm_setr_epi8(9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 1)),
-        state.has_error);
-  } else {
-    // it is not ascii so we have to do heavy work
-    state.previous = checkUTF8Bytes(in.v0, &(state.previous), &(state.has_error));
-    state.previous = checkUTF8Bytes(in.v1, &(state.previous), &(state.has_error));
-  }
-
-  if ((_mm_testz_si128(_mm_or_si128(in.v2, in.v3), highbit)) == 1) {
-    // it is ascii, we just check continuation
-    state.has_error = _mm_or_si128(
-            _mm_cmpgt_epi8(
-                    state.previous.carried_continuations,
-                    _mm_setr_epi8(9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 1)),
-            state.has_error);
-  } else {
-    // it is not ascii so we have to do heavy work
-    state.previous = checkUTF8Bytes(in.v2, &(state.previous), &(state.has_error));
-    state.previous = checkUTF8Bytes(in.v3, &(state.previous), &(state.has_error));
-  }
-}
-UNTARGET_REGION
-#endif // IS_X86_64
-
-#ifdef IS_ARM64
-template<> really_inline
-void check_utf8<architecture::arm64>(simd_input<architecture::arm64> in,
-                utf8_checking_state<architecture::arm64>& state) {
-  if (check_ascii_neon(in)) {
-    // All bytes are ascii. Therefore the byte that was just before must be ascii too.
-    // We only check the byte that was just before simd_input. Nines are arbitrary values.
-    const int8x16_t verror = (int8x16_t){9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 1};
-    state.has_error =
-        vorrq_s8(vreinterpretq_s8_u8(vcgtq_s8(state.previous.carried_continuations,
-                                    verror)),
-                     state.has_error);
-  } else {
-    // it is not ascii so we have to do heavy work
-    state.previous = checkUTF8Bytes(vreinterpretq_s8_u8(in.i0), &(state.previous), &(state.has_error));
-    state.previous = checkUTF8Bytes(vreinterpretq_s8_u8(in.i1), &(state.previous), &(state.has_error));
-    state.previous = checkUTF8Bytes(vreinterpretq_s8_u8(in.i2), &(state.previous), &(state.has_error));
-    state.previous = checkUTF8Bytes(vreinterpretq_s8_u8(in.i3), &(state.previous), &(state.has_error));
-  }
-}
-#endif // IS_ARM64
 
 // Checks if the utf8 validation has found any error.
 template<architecture T>
 errorValues check_utf8_errors(utf8_checking_state<T>& state);
 
-#ifdef IS_X86_64
-TARGET_HASWELL
-template<> really_inline
-errorValues check_utf8_errors<architecture::haswell>(utf8_checking_state<architecture::haswell>& state) {
-  return _mm256_testz_si256(state.has_error, state.has_error) == 0 ? simdjson::UTF8_ERROR : simdjson::SUCCESS;
-}
-UNTARGET_REGION
-
-TARGET_WESTMERE
-template<> really_inline
-errorValues check_utf8_errors<architecture::westmere>(utf8_checking_state<architecture::westmere>& state) {
-  return _mm_testz_si128(state.has_error, state.has_error) == 0 ? simdjson::UTF8_ERROR : simdjson::SUCCESS;
-}
-UNTARGET_REGION
-#endif
-
-#ifdef IS_ARM64
-template<> really_inline
-errorValues check_utf8_errors<architecture::arm64>(utf8_checking_state<architecture::arm64>& state) {
-  uint64x2_t v64 = vreinterpretq_u64_s8(state.has_error);
-  uint32x2_t v32 = vqmovn_u64(v64);
-  uint64x1_t result = vreinterpret_u64_u32(v32);
-  return vget_lane_u64(result, 0) != 0 ? simdjson::UTF8_ERROR : simdjson::SUCCESS;
-}
-#endif
-
-template<architecture T>
-simd_input<T> fill_input(const uint8_t * ptr);
-
-#ifdef IS_X86_64
-TARGET_HASWELL
-template<> really_inline
-simd_input<architecture::haswell> fill_input<architecture::haswell>(const uint8_t * ptr) {
-  struct simd_input<architecture::haswell> in;
-  in.lo = _mm256_loadu_si256(reinterpret_cast<const __m256i *>(ptr + 0));
-  in.hi = _mm256_loadu_si256(reinterpret_cast<const __m256i *>(ptr + 32));
-  return in;
-}
-UNTARGET_REGION
-
-TARGET_WESTMERE
-template<> really_inline
-simd_input<architecture::westmere> fill_input<architecture::westmere>(const uint8_t * ptr) {
-  struct simd_input<architecture::westmere> in;
-  in.v0 = _mm_loadu_si128(reinterpret_cast<const __m128i *>(ptr + 0));
-  in.v1 = _mm_loadu_si128(reinterpret_cast<const __m128i *>(ptr + 16));
-  in.v2 = _mm_loadu_si128(reinterpret_cast<const __m128i *>(ptr + 32));
-  in.v3 = _mm_loadu_si128(reinterpret_cast<const __m128i *>(ptr + 48));
-  return in;
-}
-UNTARGET_REGION
-#endif
-
-#ifdef IS_ARM64
-template<> really_inline
-simd_input<architecture::arm64> fill_input<architecture::arm64>(const uint8_t * ptr) {
-  struct simd_input<architecture::arm64> in;
-#ifndef TRANSPOSE
-  in.i0 = vld1q_u8(ptr + 0);
-  in.i1 = vld1q_u8(ptr + 16);
-  in.i2 = vld1q_u8(ptr + 32);
-  in.i3 = vld1q_u8(ptr + 48);
-#else
-  in.i = vld4q_u8(ptr);
-#endif
-  return in;
-}
-#endif
-
-// a straightforward comparison of a mask against input. 5 uops; would be
-// cheaper in AVX512.
+// a straightforward comparison of a mask against input. 
 template<architecture T>
 uint64_t cmp_mask_against_input(simd_input<T> in, uint8_t m);
 
-#ifdef IS_X86_64
-TARGET_HASWELL
-template<> really_inline
-uint64_t cmp_mask_against_input<architecture::haswell>(simd_input<architecture::haswell> in, uint8_t m) {
-  const __m256i mask = _mm256_set1_epi8(m);
-  __m256i cmp_res_0 = _mm256_cmpeq_epi8(in.lo, mask);
-  uint64_t res_0 = static_cast<uint32_t>(_mm256_movemask_epi8(cmp_res_0));
-  __m256i cmp_res_1 = _mm256_cmpeq_epi8(in.hi, mask);
-  uint64_t res_1 = _mm256_movemask_epi8(cmp_res_1);
-  return res_0 | (res_1 << 32);
-}
-UNTARGET_REGION
+ 
+template<architecture T>
+simd_input<T> fill_input(const uint8_t * ptr);
 
-TARGET_WESTMERE
-template<> really_inline
-uint64_t cmp_mask_against_input<architecture::westmere>(simd_input<architecture::westmere> in, uint8_t m) {
-  const __m128i mask = _mm_set1_epi8(m);
-  __m128i cmp_res_0 = _mm_cmpeq_epi8(in.v0, mask);
-  uint64_t res_0 = _mm_movemask_epi8(cmp_res_0);
-  __m128i cmp_res_1 = _mm_cmpeq_epi8(in.v1, mask);
-  uint64_t res_1 = _mm_movemask_epi8(cmp_res_1);
-  __m128i cmp_res_2 = _mm_cmpeq_epi8(in.v2, mask);
-  uint64_t res_2 = _mm_movemask_epi8(cmp_res_2);
-  __m128i cmp_res_3 = _mm_cmpeq_epi8(in.v3, mask);
-  uint64_t res_3 = _mm_movemask_epi8(cmp_res_3);
-  return res_0 | (res_1 << 16) | (res_2 << 32) | (res_3 << 48);
-}
-UNTARGET_REGION
-#endif
-
-#ifdef IS_ARM64
-template<> really_inline
-uint64_t cmp_mask_against_input<architecture::arm64>(simd_input<architecture::arm64> in, uint8_t m) {
-  const uint8x16_t mask = vmovq_n_u8(m); 
-  uint8x16_t cmp_res_0 = vceqq_u8(in.i0, mask); 
-  uint8x16_t cmp_res_1 = vceqq_u8(in.i1, mask); 
-  uint8x16_t cmp_res_2 = vceqq_u8(in.i2, mask); 
-  uint8x16_t cmp_res_3 = vceqq_u8(in.i3, mask); 
-  return neonmovemask_bulk(cmp_res_0, cmp_res_1, cmp_res_2, cmp_res_3);
-}
-#endif
 
 // find all values less than or equal than the content of maxval (using unsigned arithmetic) 
 template<architecture T>
 uint64_t unsigned_lteq_against_input(simd_input<T> in, uint8_t m);
 
-#ifdef IS_X86_64
-TARGET_HASWELL
-template<> really_inline
-uint64_t unsigned_lteq_against_input<architecture::haswell>(simd_input<architecture::haswell> in, uint8_t m) {
-  const __m256i maxval = _mm256_set1_epi8(m);
-  __m256i cmp_res_0 = _mm256_cmpeq_epi8(_mm256_max_epu8(maxval,in.lo),maxval);
-  uint64_t res_0 = static_cast<uint32_t>(_mm256_movemask_epi8(cmp_res_0));
-  __m256i cmp_res_1 = _mm256_cmpeq_epi8(_mm256_max_epu8(maxval,in.hi),maxval);
-  uint64_t res_1 = _mm256_movemask_epi8(cmp_res_1);
-  return res_0 | (res_1 << 32);
-}
-UNTARGET_REGION
-
-TARGET_WESTMERE
-template<> really_inline
-uint64_t unsigned_lteq_against_input<architecture::westmere>(simd_input<architecture::westmere> in, uint8_t m) {
-  const __m128i maxval = _mm_set1_epi8(m);
-  __m128i cmp_res_0 = _mm_cmpeq_epi8(_mm_max_epu8(maxval,in.v0),maxval);
-  uint64_t res_0 = _mm_movemask_epi8(cmp_res_0);
-  __m128i cmp_res_1 = _mm_cmpeq_epi8(_mm_max_epu8(maxval,in.v1),maxval);
-  uint64_t res_1 = _mm_movemask_epi8(cmp_res_1);
-  __m128i cmp_res_2 = _mm_cmpeq_epi8(_mm_max_epu8(maxval,in.v2),maxval);
-  uint64_t res_2 = _mm_movemask_epi8(cmp_res_2);
-  __m128i cmp_res_3 = _mm_cmpeq_epi8(_mm_max_epu8(maxval,in.v3),maxval);
-  uint64_t res_3 = _mm_movemask_epi8(cmp_res_3);
-  return res_0 | (res_1 << 16) | (res_2 << 32) | (res_3 << 48);
-}
-UNTARGET_REGION
-#endif
-
-#ifdef IS_ARM64
-template<> really_inline
-uint64_t unsigned_lteq_against_input<architecture::arm64>(simd_input<architecture::arm64> in, uint8_t m) {
-  const uint8x16_t mask = vmovq_n_u8(m); 
-  uint8x16_t cmp_res_0 = vcleq_u8(in.i0, mask); 
-  uint8x16_t cmp_res_1 = vcleq_u8(in.i1, mask); 
-  uint8x16_t cmp_res_2 = vcleq_u8(in.i2, mask); 
-  uint8x16_t cmp_res_3 = vcleq_u8(in.i3, mask); 
-  return neonmovemask_bulk(cmp_res_0, cmp_res_1, cmp_res_2, cmp_res_3);
-}
-#endif
-
-// return a bitvector indicating where we have characters that end an odd-length
-// sequence of backslashes (and thus change the behavior of the next character
-// to follow). A even-length sequence of backslashes, and, for that matter, the
-// largest even-length prefix of our odd-length sequence of backslashes, simply
-// modify the behavior of the backslashes themselves.
-// We also update the prev_iter_ends_odd_backslash reference parameter to
-// indicate whether we end an iteration on an odd-length sequence of
-// backslashes, which modifies our subsequent search for odd-length
-// sequences of backslashes in an obvious way.
-// We need to compile that code for multiple architectures. However, target attributes can be used
-// only once by function definition. Huge macro seemed better than huge code duplication.
-// uint64_t FIND_ODD_BACKSLASH_SEQUENCES(architecture T, simd_input<T> in, uint64_t &prev_iter_ends_odd_backslash)
-#define FIND_ODD_BACKSLASH_SEQUENCES(T, in, prev_iter_ends_odd_backslash) {     \
-  const uint64_t even_bits = 0x5555555555555555ULL;                             \
-  const uint64_t odd_bits = ~even_bits;                                         \
-  uint64_t bs_bits = cmp_mask_against_input<T>(in, '\\');                       \
-  uint64_t start_edges = bs_bits & ~(bs_bits << 1);                             \
-  /* flip lowest if we have an odd-length run at the end of the prior        */ \
-  /* iteration                                                               */ \
-  uint64_t even_start_mask = even_bits ^ prev_iter_ends_odd_backslash;          \
-  uint64_t even_starts = start_edges & even_start_mask;                         \
-  uint64_t odd_starts = start_edges & ~even_start_mask;                         \
-  uint64_t even_carries = bs_bits + even_starts;                                \
-                                                                                \
-  uint64_t odd_carries;                                                         \
-  /* must record the carry-out of our odd-carries out of bit 63; this        */ \
-  /* indicates whether the sense of any edge going to the next iteration     */ \
-  /* should be flipped                                                       */ \
-  bool iter_ends_odd_backslash =                                                \
-      add_overflow(bs_bits, odd_starts, &odd_carries);                          \
-                                                                                \
-  odd_carries |=                                                                \
-      prev_iter_ends_odd_backslash;  /* push in bit zero as a potential end  */ \
-                                     /* if we had an odd-numbered run at the */ \
-                                     /* end of the previous iteration        */ \
-  prev_iter_ends_odd_backslash = iter_ends_odd_backslash ? 0x1ULL : 0x0ULL;     \
-  uint64_t even_carry_ends = even_carries & ~bs_bits;                           \
-  uint64_t odd_carry_ends = odd_carries & ~bs_bits;                             \
-  uint64_t even_start_odd_end = even_carry_ends & odd_bits;                     \
-  uint64_t odd_start_even_end = odd_carry_ends & even_bits;                     \
-  uint64_t odd_ends = even_start_odd_end | odd_start_even_end;                  \
-  return odd_ends;                                                              \
-}                                                                               \
 
 template<architecture T> really_inline
 uint64_t find_odd_backslash_sequences(simd_input<T> in, uint64_t &prev_iter_ends_odd_backslash);
 
-#ifdef IS_X86_64
-TARGET_HASWELL
-template<> really_inline
-uint64_t find_odd_backslash_sequences<architecture::haswell>(simd_input<architecture::haswell> in, uint64_t &prev_iter_ends_odd_backslash) {
-  FIND_ODD_BACKSLASH_SEQUENCES(architecture::haswell, in, prev_iter_ends_odd_backslash);
-}
-UNTARGET_REGION
-
-TARGET_WESTMERE
-template<> really_inline
-uint64_t find_odd_backslash_sequences<architecture::westmere>(simd_input<architecture::westmere> in, uint64_t &prev_iter_ends_odd_backslash) {
-  FIND_ODD_BACKSLASH_SEQUENCES(architecture::westmere, in, prev_iter_ends_odd_backslash);
-}
-UNTARGET_REGION
-#endif // IS_X86_64
-
-#ifdef ARM64
-template<> really_inline
-uint64_t find_odd_backslash_sequences<architecture::arm64>(simd_input<architecture::arm64> in, uint64_t &prev_iter_ends_odd_backslash) {
-  FIND_ODD_BACKSLASH_SEQUENCES(architecture::arm64, in, prev_iter_ends_odd_backslash);
-}
-#endif // ARM64
-
-// return both the quote mask (which is a half-open mask that covers the first
-// quote
-// in an unescaped quote pair and everything in the quote pair) and the quote
-// bits, which are the simple
-// unescaped quoted bits. We also update the prev_iter_inside_quote value to
-// tell the next iteration
-// whether we finished the final iteration inside a quote pair; if so, this
-// inverts our behavior of
-// whether we're inside quotes for the next iteration.
-// Note that we don't do any error checking to see if we have backslash
-// sequences outside quotes; these
-// backslash sequences (of any length) will be detected elsewhere.
-// We need to compile that code for multiple architectures. However, target attributes can be used
-// only once by function definition. Huge macro seemed better than huge code duplication.
-// uint64_t FIND_QUOTE_MASK_AND_BITS(architecture T, simd_input<T> in, uint64_t odd_ends,
-//    uint64_t &prev_iter_inside_quote, uint64_t &quote_bits, uint64_t &error_mask)
-#define FIND_QUOTE_MASK_AND_BITS(T, in, odd_ends, prev_iter_inside_quote, quote_bits, error_mask) {    \
-  quote_bits = cmp_mask_against_input<T>(in, '"');                                                     \
-  quote_bits = quote_bits & ~odd_ends;                                                                 \
-  uint64_t quote_mask = compute_quote_mask<T>(quote_bits);                                             \
-  quote_mask ^= prev_iter_inside_quote;                                                                \
-  /* All Unicode characters may be placed within the                                                */ \
-  /* quotation marks, except for the characters that MUST be escaped:                               */ \
-  /* quotation mark, reverse solidus, and the control characters (U+0000                            */ \
-  /*through U+001F).                                                                                */ \
-  /* https://tools.ietf.org/html/rfc8259                                                            */ \
-  uint64_t unescaped = unsigned_lteq_against_input<T>(in, 0x1F);                                       \
-  error_mask |= quote_mask & unescaped;                                                                \
-  /* right shift of a signed value expected to be well-defined and standard                         */ \
-  /* compliant as of C++20,                                                                         */ \
-  /* John Regher from Utah U. says this is fine code                                                */ \
-  prev_iter_inside_quote =                                                                             \
-      static_cast<uint64_t>(static_cast<int64_t>(quote_mask) >> 63);                                   \
-  return quote_mask;                                                                                   \
-}                                                                                                      \
 
 template<architecture T> really_inline
 uint64_t find_quote_mask_and_bits(simd_input<T> in, uint64_t odd_ends,
     uint64_t &prev_iter_inside_quote, uint64_t &quote_bits, uint64_t &error_mask);
 
-#ifdef IS_X86_64
-TARGET_HASWELL
-template<> really_inline
-uint64_t find_quote_mask_and_bits<architecture::haswell>(simd_input<architecture::haswell> in, uint64_t odd_ends,
-    uint64_t &prev_iter_inside_quote, uint64_t &quote_bits, uint64_t &error_mask) {
-  FIND_QUOTE_MASK_AND_BITS(architecture::haswell, in, odd_ends, prev_iter_inside_quote, quote_bits, error_mask)
-}
-UNTARGET_REGION
-
-TARGET_WESTMERE
-template<> really_inline
-uint64_t find_quote_mask_and_bits<architecture::westmere>(simd_input<architecture::westmere> in, uint64_t odd_ends,
-    uint64_t &prev_iter_inside_quote, uint64_t &quote_bits, uint64_t &error_mask) {
-  FIND_QUOTE_MASK_AND_BITS(architecture::westmere, in, odd_ends, prev_iter_inside_quote, quote_bits, error_mask)
-}
-UNTARGET_REGION
-#endif // IS_X86_64
-
-#ifdef IS_ARM64
-template<> really_inline
-uint64_t find_quote_mask_and_bits<architecture::arm64>(simd_input<architecture::arm64> in, uint64_t odd_ends,
-    uint64_t &prev_iter_inside_quote, uint64_t &quote_bits, uint64_t &error_mask) {
-  FIND_QUOTE_MASK_AND_BITS(architecture::arm64, in, odd_ends, prev_iter_inside_quote, quote_bits, error_mask)
-}
-#endif // IS_ARM64
 
 // do a 'shufti' to detect structural JSON characters
 // they are { 0x7b } 0x7d : 0x3a [ 0x5b ] 0x5d , 0x2c
@@ -614,282 +74,6 @@ template<architecture T>
 void find_whitespace_and_structurals(simd_input<T> in,
                                      uint64_t &whitespace,
                                      uint64_t &structurals);
-
-#ifdef IS_X86_64
-TARGET_HASWELL
-template<> really_inline
-void find_whitespace_and_structurals<architecture::haswell>(simd_input<architecture::haswell> in,
-                                                     uint64_t &whitespace,
-                                                     uint64_t &structurals) {
-#ifdef SIMDJSON_NAIVE_STRUCTURAL
-  // You should never need this naive approach, but it can be useful
-  // for research purposes
-  const __m256i mask_open_brace = _mm256_set1_epi8(0x7b);
-  __m256i struct_lo = _mm256_cmpeq_epi8(in.lo, mask_open_brace);
-  __m256i struct_hi = _mm256_cmpeq_epi8(in.hi, mask_open_brace);
-  const __m256i mask_close_brace = _mm256_set1_epi8(0x7d);
-  struct_lo = _mm256_or_si256(struct_lo,_mm256_cmpeq_epi8(in.lo, mask_close_brace));
-  struct_hi = _mm256_or_si256(struct_hi,_mm256_cmpeq_epi8(in.hi, mask_close_brace));
-  const __m256i mask_open_bracket = _mm256_set1_epi8(0x5b);
-  struct_lo = _mm256_or_si256(struct_lo,_mm256_cmpeq_epi8(in.lo, mask_open_bracket));
-  struct_hi = _mm256_or_si256(struct_hi,_mm256_cmpeq_epi8(in.hi, mask_open_bracket));
-  const __m256i mask_close_bracket = _mm256_set1_epi8(0x5d);
-  struct_lo = _mm256_or_si256(struct_lo,_mm256_cmpeq_epi8(in.lo, mask_close_bracket));
-  struct_hi = _mm256_or_si256(struct_hi,_mm256_cmpeq_epi8(in.hi, mask_close_bracket));
-  const __m256i mask_column = _mm256_set1_epi8(0x3a);
-  struct_lo = _mm256_or_si256(struct_lo,_mm256_cmpeq_epi8(in.lo, mask_column));
-  struct_hi = _mm256_or_si256(struct_hi,_mm256_cmpeq_epi8(in.hi, mask_column));
-  const __m256i mask_comma = _mm256_set1_epi8(0x2c);
-  struct_lo = _mm256_or_si256(struct_lo,_mm256_cmpeq_epi8(in.lo, mask_comma));
-  struct_hi = _mm256_or_si256(struct_hi,_mm256_cmpeq_epi8(in.hi, mask_comma));
-  uint64_t structural_res_0 = static_cast<uint32_t>(_mm256_movemask_epi8(struct_lo));
-  uint64_t structural_res_1 = _mm256_movemask_epi8(struct_hi);
-  structurals = (structural_res_0 | (structural_res_1 << 32));
-
-  const __m256i mask_space = _mm256_set1_epi8(0x20);
-  __m256i space_lo = _mm256_cmpeq_epi8(in.lo, mask_space);
-  __m256i space_hi = _mm256_cmpeq_epi8(in.hi, mask_space);
-  const __m256i mask_linefeed = _mm256_set1_epi8(0x0a);
-  space_lo = _mm256_or_si256(space_lo,_mm256_cmpeq_epi8(in.lo, mask_linefeed));
-  space_hi = _mm256_or_si256(space_hi,_mm256_cmpeq_epi8(in.hi, mask_linefeed));
-  const __m256i mask_tab = _mm256_set1_epi8(0x09);
-  space_lo = _mm256_or_si256(space_lo,_mm256_cmpeq_epi8(in.lo, mask_tab));
-  space_hi = _mm256_or_si256(space_hi,_mm256_cmpeq_epi8(in.hi, mask_tab));
-  const __m256i mask_carriage = _mm256_set1_epi8(0x0d);
-  space_lo = _mm256_or_si256(space_lo,_mm256_cmpeq_epi8(in.lo, mask_carriage));
-  space_hi = _mm256_or_si256(space_hi,_mm256_cmpeq_epi8(in.hi, mask_carriage));
-
-  uint64_t ws_res_0 = static_cast<uint32_t>(_mm256_movemask_epi8(space_lo));
-  uint64_t ws_res_1 = _mm256_movemask_epi8(space_hi);
-  whitespace = (ws_res_0 | (ws_res_1 << 32));
-  // end of naive approach
-
-#else // SIMDJSON_NAIVE_STRUCTURAL
-  const __m256i structural_table = _mm256_setr_epi8(
-      44, 125, 0, 0, 0xc0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 58, 123, 
-      44, 125, 0, 0, 0xc0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 58, 123);
-  const __m256i white_table = _mm256_setr_epi8(
-      32,  100,  100,  100,  17,  100,  113,  2,  100,  9,  10,  112,  100,  13,  100,  100, 
-      32,  100,  100,  100,  17,  100,  113,  2,  100,  9,  10,  112,  100,  13,  100,  100);
-  const __m256i struct_offset = _mm256_set1_epi8(0xd4);
-  const __m256i struct_mask = _mm256_set1_epi8(32);
-
-  __m256i lo_white = _mm256_cmpeq_epi8(in.lo, 
-           _mm256_shuffle_epi8(white_table, in.lo));
-  __m256i hi_white = _mm256_cmpeq_epi8(in.hi, 
-           _mm256_shuffle_epi8(white_table, in.hi));
-  uint64_t ws_res_0 = static_cast<uint32_t>(_mm256_movemask_epi8(lo_white));
-  uint64_t ws_res_1 = _mm256_movemask_epi8(hi_white);
-  whitespace = (ws_res_0 | (ws_res_1 << 32));
-  __m256i lo_struct_r1 = _mm256_add_epi8(struct_offset, in.lo);
-  __m256i hi_struct_r1 = _mm256_add_epi8(struct_offset, in.hi);
-  __m256i lo_struct_r2 = _mm256_or_si256(in.lo, struct_mask);
-  __m256i hi_struct_r2 = _mm256_or_si256(in.hi, struct_mask);
-  __m256i lo_struct_r3 = _mm256_shuffle_epi8(structural_table, lo_struct_r1);
-  __m256i hi_struct_r3 = _mm256_shuffle_epi8(structural_table, hi_struct_r1);
-  __m256i lo_struct = _mm256_cmpeq_epi8(lo_struct_r2, lo_struct_r3);
-  __m256i hi_struct = _mm256_cmpeq_epi8(hi_struct_r2, hi_struct_r3);
-  
-  uint64_t structural_res_0 =
-      static_cast<uint32_t>(_mm256_movemask_epi8(lo_struct));
-  uint64_t structural_res_1 = _mm256_movemask_epi8(hi_struct);
-  structurals = (structural_res_0 | (structural_res_1 << 32));
-#endif // SIMDJSON_NAIVE_STRUCTURAL
-}
-UNTARGET_REGION
-
-TARGET_WESTMERE
-template<> really_inline
-void find_whitespace_and_structurals<architecture::westmere>(simd_input<architecture::westmere> in,
-                                                     uint64_t &whitespace, uint64_t &structurals) {
-  const __m128i structural_table = _mm_setr_epi8(44, 125, 0, 0, 0xc0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 58, 123);
-  const __m128i white_table = _mm_setr_epi8(
-      32,  100,  100,  100,  17,  100,  113,  2,  100,  9,  10,  112,  100,  13,  100,  100);
-  const __m128i struct_offset = _mm_set1_epi8(0xd4);
-  const __m128i struct_mask = _mm_set1_epi8(32);
-
-  __m128i white0 = _mm_cmpeq_epi8(in.v0,
-           _mm_shuffle_epi8(white_table, in.v0));
-  __m128i white1 = _mm_cmpeq_epi8(in.v1,
-           _mm_shuffle_epi8(white_table, in.v1));
-  __m128i white2 = _mm_cmpeq_epi8(in.v2,
-           _mm_shuffle_epi8(white_table, in.v2));
-  __m128i white3 = _mm_cmpeq_epi8(in.v3,
-           _mm_shuffle_epi8(white_table, in.v3));
-  uint64_t ws_res_0 = _mm_movemask_epi8(white0);
-  uint64_t ws_res_1 = _mm_movemask_epi8(white1);
-  uint64_t ws_res_2 = _mm_movemask_epi8(white2);
-  uint64_t ws_res_3 = _mm_movemask_epi8(white3);
-
-  whitespace = (ws_res_0 | (ws_res_1 << 16) | (ws_res_2 << 32) | (ws_res_3 << 48));
-
-  __m128i struct1_r1 = _mm_add_epi8(struct_offset, in.v0);
-  __m128i struct2_r1 = _mm_add_epi8(struct_offset, in.v1);
-  __m128i struct3_r1 = _mm_add_epi8(struct_offset, in.v2);
-  __m128i struct4_r1 = _mm_add_epi8(struct_offset, in.v3);
-
-  __m128i struct1_r2 = _mm_or_si128(in.v0, struct_mask);
-  __m128i struct2_r2 = _mm_or_si128(in.v1, struct_mask);
-  __m128i struct3_r2 = _mm_or_si128(in.v2, struct_mask);
-  __m128i struct4_r2 = _mm_or_si128(in.v3, struct_mask);
-
-  __m128i struct1_r3 = _mm_shuffle_epi8(structural_table, struct1_r1);
-  __m128i struct2_r3 = _mm_shuffle_epi8(structural_table, struct2_r1);
-  __m128i struct3_r3 = _mm_shuffle_epi8(structural_table, struct3_r1);
-  __m128i struct4_r3 = _mm_shuffle_epi8(structural_table, struct4_r1);
-
-  __m128i struct1 = _mm_cmpeq_epi8(struct1_r2, struct1_r3);
-  __m128i struct2 = _mm_cmpeq_epi8(struct2_r2, struct2_r3);
-  __m128i struct3 = _mm_cmpeq_epi8(struct3_r2, struct3_r3);
-  __m128i struct4 = _mm_cmpeq_epi8(struct4_r2, struct4_r3);
-
-  uint64_t structural_res_0 = _mm_movemask_epi8(struct1);
-  uint64_t structural_res_1 = _mm_movemask_epi8(struct2);
-  uint64_t structural_res_2 = _mm_movemask_epi8(struct3);
-  uint64_t structural_res_3 = _mm_movemask_epi8(struct4);
-
-  structurals = (structural_res_0 | (structural_res_1 << 16) | (structural_res_2 << 32) | (structural_res_3 << 48));
-}
-UNTARGET_REGION
-#endif // X86_64
-
-#ifdef IS_ARM64
-template<> really_inline
-void find_whitespace_and_structurals<architecture::arm64>(
-                                                  simd_input<architecture::arm64> in,
-                                                  uint64_t &whitespace,
-                                                  uint64_t &structurals) {
-  const uint8x16_t low_nibble_mask = (uint8x16_t){ 
-      16, 0, 0, 0, 0, 0, 0, 0, 0, 8, 12, 1, 2, 9, 0, 0};
-  const uint8x16_t high_nibble_mask = (uint8x16_t){ 
-      8, 0, 18, 4, 0, 1, 0, 1, 0, 0, 0, 3, 2, 1, 0, 0};
-  const uint8x16_t structural_shufti_mask = vmovq_n_u8(0x7); 
-  const uint8x16_t whitespace_shufti_mask = vmovq_n_u8(0x18); 
-  const uint8x16_t low_nib_and_mask = vmovq_n_u8(0xf); 
-
-  uint8x16_t nib_0_lo = vandq_u8(in.i0, low_nib_and_mask);
-  uint8x16_t nib_0_hi = vshrq_n_u8(in.i0, 4);
-  uint8x16_t shuf_0_lo = vqtbl1q_u8(low_nibble_mask, nib_0_lo);
-  uint8x16_t shuf_0_hi = vqtbl1q_u8(high_nibble_mask, nib_0_hi);
-  uint8x16_t v_0 = vandq_u8(shuf_0_lo, shuf_0_hi);
-
-  uint8x16_t nib_1_lo = vandq_u8(in.i1, low_nib_and_mask);
-  uint8x16_t nib_1_hi = vshrq_n_u8(in.i1, 4);
-  uint8x16_t shuf_1_lo = vqtbl1q_u8(low_nibble_mask, nib_1_lo);
-  uint8x16_t shuf_1_hi = vqtbl1q_u8(high_nibble_mask, nib_1_hi);
-  uint8x16_t v_1 = vandq_u8(shuf_1_lo, shuf_1_hi);
-
-  uint8x16_t nib_2_lo = vandq_u8(in.i2, low_nib_and_mask);
-  uint8x16_t nib_2_hi = vshrq_n_u8(in.i2, 4);
-  uint8x16_t shuf_2_lo = vqtbl1q_u8(low_nibble_mask, nib_2_lo);
-  uint8x16_t shuf_2_hi = vqtbl1q_u8(high_nibble_mask, nib_2_hi);
-  uint8x16_t v_2 = vandq_u8(shuf_2_lo, shuf_2_hi);
-
-  uint8x16_t nib_3_lo = vandq_u8(in.i3, low_nib_and_mask);
-  uint8x16_t nib_3_hi = vshrq_n_u8(in.i3, 4);
-  uint8x16_t shuf_3_lo = vqtbl1q_u8(low_nibble_mask, nib_3_lo);
-  uint8x16_t shuf_3_hi = vqtbl1q_u8(high_nibble_mask, nib_3_hi);
-  uint8x16_t v_3 = vandq_u8(shuf_3_lo, shuf_3_hi);
-
-  uint8x16_t tmp_0 = vtstq_u8(v_0, structural_shufti_mask);
-  uint8x16_t tmp_1 = vtstq_u8(v_1, structural_shufti_mask);
-  uint8x16_t tmp_2 = vtstq_u8(v_2, structural_shufti_mask);
-  uint8x16_t tmp_3 = vtstq_u8(v_3, structural_shufti_mask);
-  structurals = neonmovemask_bulk(tmp_0, tmp_1, tmp_2, tmp_3);
-
-  uint8x16_t tmp_ws_0 = vtstq_u8(v_0, whitespace_shufti_mask);
-  uint8x16_t tmp_ws_1 = vtstq_u8(v_1, whitespace_shufti_mask);
-  uint8x16_t tmp_ws_2 = vtstq_u8(v_2, whitespace_shufti_mask);
-  uint8x16_t tmp_ws_3 = vtstq_u8(v_3, whitespace_shufti_mask);
-  whitespace = neonmovemask_bulk(tmp_ws_0, tmp_ws_1, tmp_ws_2, tmp_ws_3);
-}
-#endif // __ARM_NEON
-
-
-#ifdef SIMDJSON_NAIVE_FLATTEN // useful for benchmarking
-//
-// This is just a naive implementation. It should be normally
-// disable, but can be used for research purposes to compare
-// again our optimized version.
-really_inline void flatten_bits(uint32_t *base_ptr, uint32_t &base,
-                                uint32_t idx, uint64_t bits) {
-  uint32_t * out_ptr = base_ptr + base;
-  idx -= 64;
-  while(bits != 0) {
-      out_ptr[0] = idx + trailingzeroes(bits);
-      bits = bits & (bits - 1);
-      out_ptr++;
-  }
-  base = (out_ptr - base_ptr);
-}
-
-#else 
-// flatten out values in 'bits' assuming that they are are to have values of idx
-// plus their position in the bitvector, and store these indexes at
-// base_ptr[base] incrementing base as we go
-// will potentially store extra values beyond end of valid bits, so base_ptr
-// needs to be large enough to handle this
-really_inline void flatten_bits(uint32_t *base_ptr, uint32_t &base,
-                                uint32_t idx, uint64_t bits) {
-  // In some instances, the next branch is expensive because it is mispredicted. 
-  // Unfortunately, in other cases,
-  // it helps tremendously.
-  if(bits == 0) return; 
-  uint32_t cnt = hamming(bits);
-  uint32_t next_base = base + cnt;
-  idx -= 64;
-  base_ptr += base;
-  { 
-    base_ptr[0] = idx + trailingzeroes(bits);
-    bits = bits & (bits - 1);
-    base_ptr[1] = idx + trailingzeroes(bits);
-    bits = bits & (bits - 1);
-    base_ptr[2] = idx + trailingzeroes(bits);
-    bits = bits & (bits - 1);
-    base_ptr[3] = idx + trailingzeroes(bits);
-    bits = bits & (bits - 1);
-    base_ptr[4] = idx + trailingzeroes(bits);
-    bits = bits & (bits - 1);
-    base_ptr[5] = idx + trailingzeroes(bits);
-    bits = bits & (bits - 1);
-    base_ptr[6] = idx + trailingzeroes(bits);
-    bits = bits & (bits - 1);
-    base_ptr[7] = idx + trailingzeroes(bits);
-    bits = bits & (bits - 1);
-    base_ptr += 8;
-  }
-  // We hope that the next branch is easily predicted.
-  if (cnt > 8) {
-    base_ptr[0] = idx + trailingzeroes(bits);
-    bits = bits & (bits - 1);
-    base_ptr[1] = idx + trailingzeroes(bits);
-    bits = bits & (bits - 1);
-    base_ptr[2] = idx + trailingzeroes(bits);
-    bits = bits & (bits - 1);
-    base_ptr[3] = idx + trailingzeroes(bits);
-    bits = bits & (bits - 1);
-    base_ptr[4] = idx + trailingzeroes(bits);
-    bits = bits & (bits - 1);
-    base_ptr[5] = idx + trailingzeroes(bits);
-    bits = bits & (bits - 1);
-    base_ptr[6] = idx + trailingzeroes(bits);
-    bits = bits & (bits - 1);
-    base_ptr[7] = idx + trailingzeroes(bits);
-    bits = bits & (bits - 1);
-    base_ptr += 8;
-  }
-  if (cnt > 16) { // unluckly: we rarely get here
-    // since it means having one structural or pseudo-structral element 
-    // every 4 characters (possible with inputs like "","","",...).
-    do {
-      base_ptr[0] = idx + trailingzeroes(bits);
-      bits = bits & (bits - 1);
-      base_ptr++;
-    } while(bits != 0);
-  }
-  base = next_base;
-}
-#endif // SIMDJSON_NAIVE_FLATTEN
 
 // return a updated structural bit vector with quoted contents cleared out and
 // pseudo-structural characters added to the mask
@@ -931,13 +115,15 @@ really_inline uint64_t finalize_structurals(
 }
 
 template<architecture T = architecture::native>
-/*never_inline*/ int find_structural_bits(const uint8_t *buf, size_t len, ParsedJson &pj);
+int find_structural_bits(const uint8_t *buf, size_t len, simdjson::ParsedJson &pj);
 
 
-template<architecture T = architecture::native>
-WARN_UNUSED
-int find_structural_bits(const char *buf, size_t len, ParsedJson &pj) {
-  return find_structural_bits<T>(reinterpret_cast<const uint8_t *>(buf), len, pj);
-}
-}
+} // namespace simdjson
+
+
+
+
+
+
+
 #endif
